python/algorithms/igraph_cmp/requirements.txt
pytest==6.2.4
google-benchmark==0.1.2
black
python-igraph==0.9.6
